feat(#1706): refactored
