+++ apps/web/src/app/[lang]/dho/[id]/@aside/agreements/create/change-voting-method/page.tsx
@@ -5,7 +5,7 @@ import {
 import { Locale } from '@hypha-platform/i18n';
 import { notFound } from 'next/navigation';
 import { getDhoPathAgreements } from '../../../../@tab/agreements/constants';
-import { Plugin } from '../plugins';
+import { Plugin } from '../../../../_components/plugins';
 import { PATH_SELECT_SETTINGS_ACTION } from '@web/app/constants';
 import { findSpaceBySlug } from '@hypha-platform/core/server';
 import { db } from '@hypha-platform/storage-postgres';
