+++ apps/web/src/app/[lang]/dho/[id]/@aside/agreements/create/deploy-funds/page.tsx
@@ -1,7 +1,7 @@
 import { CreateDeployFundsForm, SidePanel } from '@hypha-platform/epics';
 import { Locale } from '@hypha-platform/i18n';
 import { getDhoPathAgreements } from '../../../../@tab/agreements/constants';
-import { Plugin } from '../plugins';
+import { Plugin } from '../../../../_components/plugins';
 import { notFound } from 'next/navigation';
 import { PATH_SELECT_CREATE_ACTION } from '@web/app/constants';
 import { findSpaceBySlug } from '@hypha-platform/core/server';
