+++ apps/web/src/app/[lang]/dho/[id]/@aside/agreements/create/propose-contribution/page.tsx
@@ -4,7 +4,7 @@ import {
 } from '@hypha-platform/epics';
 import { Locale } from '@hypha-platform/i18n';
 import { getDhoPathAgreements } from '../../../../@tab/agreements/constants';
-import { Plugin } from '../plugins';
+import { Plugin } from '../../../../_components/plugins';
 import { notFound } from 'next/navigation';
 import { PATH_SELECT_CREATE_ACTION } from '@web/app/constants';
 import { findSpaceBySlug, getAllSpaces } from '@hypha-platform/core/server';
