+++ apps/web/src/app/[lang]/dho/[id]/@aside/agreements/create/space-to-space-membership/page.tsx
@@ -8,7 +8,7 @@ import {
   type Space,
   findSpaceBySlug,
 } from '@hypha-platform/core/server';
-import { Plugin } from '../plugins';
+import { Plugin } from '../../../../_components/plugins';
 import { db } from '@hypha-platform/storage-postgres';
 import { notFound } from 'next/navigation';
