+++ apps/web/src/app/[lang]/dho/[id]/_components/plugins.tsx
